fix: static images moved from cloud to local
