+++ client/src/components/ui/collections/collection-card.tsx
@@ -2,7 +2,7 @@ import { InfoCircledIcon, LockClosedIcon } from '@radix-ui/react-icons'
 import { Link } from '@tanstack/react-router'
 import { LikeButton } from '../like-button'
 import { Collection, User } from '@/schemas/dbSchemas'
-import { dummyCollectionImage } from '@/constants/media'
+import dummyCollectionImage from '@/assets/images/dummyCollectionImage.jpg'
 import useAuthUser from 'react-auth-kit/hooks/useAuthUser'
 import { useLike } from '@/hooks/useLike'
 import ReactMarkdown from 'react-markdown'
