--- client/src/constants/media.ts
@@ -1,7 +1,0 @@
-export const dummyCollectionImage =
-  'https://firebasestorage.googleapis.com/v0/b/my-first-projecvfs324.appspot.com/o/images%2FdummyCollectionImage.jpg?alt=media&token=2b18cc20-3a7b-40ce-8474-341133db482c'
-
-export const dummyItemImage = `https://firebasestorage.googleapis.com/v0/b/my-first-projecvfs324.appspot.com/o/images%2FdummyItemImage.jpg?alt=media&token=7eca72ef-6851-4bb8-aee7-023ec965e794`
-
-export const introVideo =
-  'https://firebasestorage.googleapis.com/v0/b/my-first-projecvfs324.appspot.com/o/videos%2Fintro.mp4?alt=media&token=30fa3522-1e73-4f12-89b1-db9fff3b47ac'
